refactor: resolve #292 - split useProgramLibrary.test.ts from 600 to under 500 lines
Split into:
- useProgramLibrary.crud.test.ts (478 lines) — CRUD operations and integration
- useProgramLibrary.state.test.ts (186 lines) — state lifecycle, $reset, singleton

Closes #292
